requirements.txt
numpy >= 1.13.0
scipy
pandas
matplotlib
mpltools
pytest
pytest-timeout
sphinx
dask
distributed
gdown
flake8
torch>=2.1.0
Pillow>=9.4.0
scikit-learn
torchvision>=0.16.0
seaborn
opencv-python
pytorch-msssim